os/sshd: restructure to be more Ignition first
This commit makes several changes.

1. Suggests port 222/223 instead of 2222/2223
This is a small security thing. Because an unprivileged process can more
easily hijack unprivileged (>1024) ports, listening on those ports means
a DoS against systemd (of socket-activated)/sshd (otherwise) that stops
it from listening allows an unprivileged process to MITM the connection.

Using a lower port means that a DoS/crash does not present that
opportunity so trivially.

2. Reorders *all* CLCs to be before *all* by-hand bits
This intentionally promotes disabling socket activation to be a more
prominent section.
I think this is a good thing since it's actually a quite sane thing to
do.

3. Rewords several of the by-hand bits in an attempt to be more correct.