os/kernel-modules: do not check /etc for channel information
Previously, /etc was the only definition of the channel name.  All
recent releases write the channel to /usr now that releases are no
longer promoted, so this value is guaranteed to be correct.

This change also avoids cases of incorrect information when the
user sets a new channel in /etc before the update/reboot.