docs: Update the authentication documentation
Nebraska doesn't support groups but instead assumes that there is a
single team in the Nebraska database. This means that we need to setup
a separate Nebraska instance for separation between different customers
or projects. Update the docs accordingly.

Also, move the documentation to docs/auth.md

This commit closes #28

Signed-off-by: Ashijeet Acharya <ashijeet@kinvolk.io>