Working permanent links for releases
Anchors existed to refer to a release or
a channel but they did not work for releases
other than stable (and were used twice because
it was part of the architecture link).
Clean up the anchor definitions and introduce
a script to show the correct channel when
landing on such a link.