Note that update-engine needs to run after the reboot
The documentation explained how to disable update-engine permanently
but the fact that updates don't work if update-engine is running only
shortly to update was not called out.
Explain that update-engine needs to run after the reboot, too, and show
how to check that the service can be disabled again safely.