Comment out the section about cloud images
Our GCE docs currently include a section about deploying cloud images to
GCE that doesn't work because we don't have a public project (yet).
Comment it out for now.

We will have a public project soon, we can remove the comments then and
move the customization section to the bottom.