Make the GCP kinvolk-public project known
We had commented out the instructions for deploying GCP instances,
because they were using a non existant "flatcar-cloud" project. We now
have a "kinvolk-public" project, where we publish GCP images and we can
now publish instructions on how to install these images.

Also, the comment that images are not available is not valid anymore, so
replace that.