developer-guides/sdk-modifying-flatcar.md: use dynamic format default
The default parameter is computed based on the board: for amd64 it's
qemu but for arm64 it's qemu_uefi. For users reading the docs it may
result in generating the wrong image type for arm64 which means that
the firmware files are not copied to the target directory and have to
be added manually.

Do not hardcode the format value in the docs but omit the parameter so
that it will choose the right default automatically.