update-strategies: document how to do manual updates
The existing way of disabling automatic updates with masking the
services was difficult when the user still wanted to do manual updates
because one needs to know that unmasking needs to be kept for the first
minutes after the reboot.

Document an easier and safer way that uses the update-engine config
file to disable automatic updates.