Add guide for image customization
While users often had the need to customize the image, we had no docs
to point to. Also, there are many traps when people try to customize
by booting an image and we have to warn users about them.

Add a new document that captures the tribal knowledge of image
preparation and guides the users to select the appropriate method
without running into the traps.