Rename image building guide
The image building guide we maintain is for building from source in
contrast to a future guide of customizing a Flatcar image by modifying
it directly.

Rename the image building guide to be more clear about what it does,
in order to add a future guide for customizing Flatcar images.