switching-to-unified-cgroups: also set the bootloader kernel cmdline
The bind-mount workaround is not the recommended state the system
should be in when using cgroupv1 because it has some negative effects.

Always set the bootloader kernel cmdline explicitly as well in the
snippet example to make sure that the workaround is only needed on the
first boot. Also document one known issue.