Add LTS-<YEAR> ATOM and JSON feeds
With two LTS streams being available at the same time, the single
column or ATOM/JSON feeds are not perfect for looking up the releases
of a stream.
Provide the LTS-<YEAR> ATOM and JSON feeds for the both LTS streams to
help users track a stream or get the list of all releases of a stream.