Sort releases by version
When the LTS-2021 release gets published after the last LTS-2022
release it would show first in the list of LTS releases. Rather put
all LTS-2022 releases before all LTS-2021 releases since these are
not promoted anyway to be used and the release notes are more for
reference if one searches for them.