release-notes: Do not index into the wrong map
The global sorted map was used to get the package name through an
index operation but the index was not for the global map but came
from the local map which can have a different elements (causing the
package name to be wrong for the displayed version or the indexing
to fail).
Iterate over the local map to show the package name and version.
While not guaranteed to be sorted, in testing this was the case and
anyway, the sorting is not too critical because the entries are not
displayed besides entries of another release. I tried to sort but
it wasn't possible because the sorting replaced the map key with the
index.