hardening-guide: Document how to disable the core user
Since the group gets provided by nss-altfiles on runtime from the
read-only user DB in /usr, the groups docker and wheel can't be changed
to remove the core user.
The solution to restrict access is to disable the core user, restrict
sudo access, or to change the permissions of the Docker socket.
The rkt section can be deleted as rkt is gone.