nginx example: Use journald logging backend
The default Docker log backend is using json files and for long running
processes they could easily fill up the disk.
Use the journald logging backend to benefit from truncating. The "none"
backend is also an option but it makes "docker log ID" fail. The
drawback with the journald backend is that it duplicates the system
wide entries because we already have them from the systemd unit stdout.