sysext: Document setting up a symlink to /dev/null to disable default
The original idea was that our future Docker/containerd sysext is set
up in /usr/lib/extensions/ but support for this location got removed:
https://github.com/systemd/systemd/commit/de862276eddbbe76b436213b4d427205356d1886
Now we would have to use our /etc overlay to enable the sysext image by
default. The user opt-out entry to disable it was also in /etc and
this won't work anymore. Instead, the user should directly overwrite
the entry to let it point to /dev/null.