sysext: Document OEM extensions and Flatcar extensions
The use of systemd-sysext for OEM and Flatcar extensions was only
documented in the release notes but not yet part of the docs.
Add an overview about the systemd-sysext use in Flatcar and remove the
Butane sysupdate config because the sysext-bakery is the better place
to keep this up-to-date. This also adds initial content for ZFS docs.