docs: No need to manually delete bootengine.cpio any more
Dracut is no longer run in a chroot, so bootengine.cpio is created
alongside the other build files rather than in the board root and is
therefore no longer persisted.

To be honest, there was probably no need to delete it anyway, as it
would have been overwritten. It seems like this was just a precaution.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>