adding-swap: Fix the Butane swap file example
Perhaps it worked properly in 2017, but the create-swapfile.service
currently only works once because mkswap errors if you call it against
an existing file, even if it is not being used.

Therefore remove the swap file before trying to create it. You cannot
remove it when it is in use (contrary to Linux norms), so change
Requires to BindsTo. This ensures the swap is disabled before
create-swapfile.service (re)starts.

We could check for existence of the file instead, but swap file content
becomes irrelevant once it is disabled. It is common practise to
recreate encrypted swap partitions with a random key on every boot.

Bug: https://github.com/flatcar/Flatcar/issues/1847
Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>