Easier setup for verification
Linking to the signing key as otherwise the step to verify the images and scripts fails. And this surprises people not aware what this means.

Easier setup for verification

Linking to the signing key as otherwise the step to verify the images and scripts fails. And this surprises people not aware what this means.

End sentence correctly