kernel-modules: Fix race condition in documented systemd mount unit
The /usr/lib/modules overlay is seemingly mounted too early, preventing
the overlaid files from appearing despite the mount being listed. The
system claims that the mount does not exist when attempting to unmount.

It appears to be racing with systemd-sysext.service, so schedule it to
start after that. This happens after local-fs.target, so run before
sysinit.target instead.

Subsequently running `systemd-sysext refresh` causes the overlay to
become unmounted, but `UpheldBy` automatically mounts it again
immediately afterwards.

Applying mutable mode to /usr might be preferable, but that apparently
requires more work. This will do in the meantime.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>
Signed-off-by: PackElend <muellerstefan@gmx.net>