hetzner: update amd64 server type
The currently used server type (`cx22`) is deprecated and will be
removed on 2026-01-01. The new server type is available in the same
locations as the previous one and has the smallest root disk size, so
the generated snapshots are most compatible with other types, that is
why I did not use `cpx22` instead.

- https://www.hetzner.com/news/new-cloud-plans/
- https://docs.hetzner.cloud/changelog#2025-10-16-new-server-types-with-categories
- https://docs.hetzner.cloud/changelog#2025-10-16-server-types-deprecated

Signed-off-by: Julian Tölle <julian.toelle@hetzner-cloud.de>
Signed-off-by: PackElend <muellerstefan@gmx.net>